Refactor code: move tim and usart code into common dir
