+++ CMakeLists.txt
@@ -71,10 +71,10 @@ set(BUILD_SHARED_LIBS OFF)
 SET(STG_MODEL "model" CACHE STRING "850")
 if("${STG_MODEL}" MATCHES "850")
   message(STATUS "STG MODEL: ${STG_MODEL}")
-  add_definitions(-DMODEL_STG_850)
+  add_definitions(-DMODEL_STG850)
 elseif("${STG_MODEL}" MATCHES "856")
   message(STATUS "STG MODEL: ${STG_MODEL}")
-  add_definitions(-DMODEL_STG_856)
+  add_definitions(-DMODEL_STG856)
 elseif("${STG_MODEL}" MATCHES "TESTER")
   message(STATUS "STG MODEL: ${STG_MODEL}")
   add_definitions(-DMODEL_TESTER)
